Conflicts:
	esp/esp/program/controllers/lottery.py
	esp/esp/program/modules/handlers/lotterystudentregmodule.py

+++ esp/useful_scripts/server_setup/make_virtualenv.sh
@@ -1,15 +1,22 @@
-#!/bin/bash
+#!/bin/bash -e
 
-while [[ ! -n $BASEDIR ]]; do
+while [[ ! -n "$BASEDIR" ]]; do
     echo -n "Enter the root directory path of this site install (it should include a .git file) --> "
     read BASEDIR
 done
-FULLPATH=`mkdir -p $BASEDIR; cd $BASEDIR; pwd`
-BASEDIR=`echo "$FULLPATH" | sed -e "s/\/*$//"`
+if [[ ! -f "$BASEDIR/esp/packages.txt" || ! -f "$BASEDIR/esp/requirements.txt" ]]; then
+    echo "Couldn't find requirements files for site install at $BASEDIR"
+    exit 1
+fi
+FULLPATH=$(mkdir -p "$BASEDIR"; cd "$BASEDIR"; pwd)
+BASEDIR=$(echo "$FULLPATH" | sed -e "s/\/*$//")
 
-virtualenv $BASEDIR/env
-source $BASEDIR/env/bin/activate
-pip install -r $BASEDIR/esp/requirements.txt
+sudo apt-get install $(< "$BASEDIR/esp/packages.txt")
+
+sudo pip install "virtualenv>=1.10"
+virtualenv "$BASEDIR/env"
+source "$BASEDIR/env/bin/activate"
+pip install -r "$BASEDIR/esp/requirements.txt"
 
 cat >/tmp/esp.wsgi <<EOF
 try:
@@ -21,5 +28,5 @@ except IOError:
 
 EOF
 
-cat $BASEDIR/esp.wsgi >/tmp/esp.wsgi
-mv /tmp/esp.wsgi $BASEDIR/esp.wsgi
+cat "$BASEDIR/esp.wsgi" >>/tmp/esp.wsgi
+mv /tmp/esp.wsgi "$BASEDIR/esp.wsgi"
